made new mac project. had to because several weird bugs prevented changing AU unique ID

+++ JuceLibraryCode/modules/juce_osc/osc/juce_OSCReceiver.cpp
@@ -416,7 +416,7 @@ struct OSCReceiver::Pimpl   : private Thread,
         catch (OSCFormatError)
         {
             if (formatErrorHandler != nullptr)
-                formatErrorHandler (data, (int) dataSize);
+                formatErrorHandler (data, dataSize);
         }
     }
